DISPATCH-2003: Memory usage as uint32 underreports larger sizes

+++ src/router_core/agent_router.c
@@ -246,7 +246,7 @@ static void qdr_agent_write_column_CT(qd_composed_field_t *body, int col, qdr_co
         break;
 
     case QDR_ROUTER_MEMORY_USAGE: {
-        uint32_t size = qd_router_memory_usage();
+        uint64_t size = qd_router_memory_usage();
         if (size)
             qd_compose_insert_ulong(body, size);
         else  // memory usage not available
